+++ latch/latch/registry/table.py
@@ -1,27 +1,52 @@
 from contextlib import contextmanager
 from dataclasses import dataclass, field
+from datetime import date, datetime
+from enum import Enum
 from typing import (
     Dict,
     Iterator,
     List,
     Literal,
     Optional,
+    Tuple,
+    Type,
     TypedDict,
     Union,
     cast,
+    get_args,
+    get_origin,
     overload,
 )
 
 import gql
 import graphql.language as l
 import graphql.language.parser as lp
+from typing_extensions import TypeAlias
 
 from latch.gql._execute import execute
+from latch.gql._utils import (
+    _GqlJsonValue,
+    _json_value,
+    _name_node,
+    _parse_selection,
+    _var_def_node,
+    _var_node,
+)
 from latch.registry.record import NoSuchColumnError, Record
-from latch.registry.types import Column, InvalidValue, RecordValue, RegistryPythonValue
-from latch.registry.upstream_types.types import DBType
+from latch.registry.types import (
+    Column,
+    InvalidValue,
+    LinkedRecordType,
+    RecordValue,
+    RegistryEnumDefinition,
+    RegistryPythonType,
+    RegistryPythonValue,
+)
+from latch.registry.upstream_types.types import DBType, RegistryType
 from latch.registry.upstream_types.values import DBValue, EmptyCell
 from latch.registry.utils import to_python_literal, to_python_type, to_registry_literal
+from latch.types.directory import LatchDir
+from latch.types.file import LatchFile
 
 from ..types.json import JsonValue
 
@@ -70,7 +95,8 @@ class Table:
         Always makes a network request.
         """
         data = execute(
-            gql.gql("""
+            gql.gql(
+                """
                 query TableQuery($id: BigInt!) {
                     catalogExperiment(id: $id) {
                         id
@@ -84,7 +110,8 @@ class Table:
                         }
                     }
                 }
-                """),
+                """
+            ),
             variables={"id": self.id},
         )["catalogExperiment"]
         # todo(maximsmol): deal with nonexistent tables
@@ -182,7 +209,8 @@ class Table:
         # todo(maximsmol): because allSamples returns each column as its own
         # row, we can't paginate by samples because we don't know when a sample is finished
         nodes: List[_AllRecordsNode] = execute(
-            gql.gql("""
+            gql.gql(
+                """
                 query TableQuery($id: BigInt!) {
                     catalogExperiment(id: $id) {
                         allSamples {
@@ -195,7 +223,8 @@ class Table:
                         }
                     }
                 }
-                """),
+                """
+            ),
             {
                 "id": self.id,
             },
@@ -266,84 +295,58 @@ class Table:
         if reload_on_commit:
             self.load()
 
-    def __str__(self) -> str:
-        name = self.get_display_name(load_if_missing=False)
-        if name is not None:
-            return f"Table({repr(name)})"
+    def __repr__(self):
+        display_name = self.get_display_name(load_if_missing=False)
+        if display_name is not None:
+            return f"Table(id={self.id}, display_name={display_name})"
 
+        return f"Table(id={self.id})"
+
+    def __str__(self):
         return repr(self)
-
-
-def _parse_selection(x: str) -> l.SelectionNode:
-    p = lp.Parser(l.Source(x.lstrip()))
-    p.expect_token(l.TokenKind.SOF)
-    res = p.parse_selection()
-    p.expect_token(l.TokenKind.EOF)
-    return res
-
-
-def _name_node(x: str) -> l.NameNode:
-    res = l.NameNode()
-    res.value = x
-    return res
-
-
-def _obj_field(k: str, x: JsonValue) -> l.ObjectFieldNode:
-    res = l.ObjectFieldNode()
-
-    res.name = _name_node(k)
-    res.value = _json_value(x)
-
-    return res
-
-
-def _json_value(x: JsonValue) -> l.ValueNode:
-    # note: this does not support enums
-    if x is None:
-        return l.NullValueNode()
-
-    if isinstance(x, str):
-        res = l.StringValueNode()
-        res.value = x
-        return res
-
-    if isinstance(x, int):
-        if isinstance(x, bool):
-            res = l.BooleanValueNode()
-            res.value = x
-            return res
-
-        res = l.IntValueNode()
-        res.value = str(x)
-        return res
-
-    if isinstance(x, float):
-        res = l.FloatValueNode()
-        res.value = str(x)
-        return res
-
-    if isinstance(x, float):
-        res = l.FloatValueNode()
-        res.value = str(x)
-        return res
-
-    if isinstance(x, list):
-        res = l.ListValueNode()
-        res.values = tuple(_json_value(el) for el in x)
-        return res
-
-    if isinstance(x, dict):
-        res = l.ObjectValueNode()
-        res.fields = tuple(_obj_field(k, v) for k, v in x.items())
-        return res
-
-    raise ValueError(f"cannot Graphql-serialize JSON value of type {type(x)}: {x}")
 
 
 @dataclass(frozen=True)
 class _TableRecordsUpsertData:
     name: str
     values: Dict[str, DBValue]
+
+
+@dataclass(frozen=True)
+class _TableRecordsDeleteData:
+    name: str
+
+
+@dataclass(frozen=True)
+class _TableColumnUpsertData:
+    key: str
+    type: DBType
+
+
+_TableRecordsMutationData: TypeAlias = Union[
+    _TableRecordsUpsertData,
+    _TableRecordsDeleteData,
+    _TableColumnUpsertData,
+]
+
+
+class InvalidColumnTypeError(ValueError):
+    """Failure to use an invalid column type.
+
+    Attributes:
+        key: Identifier of the invalid column.
+        invalid_type: Requested column type.
+    """
+
+    def __init__(
+        self, key: str, invalid_type: Union[Type[object], RegistryPythonType], msg: str
+    ):
+        super().__init__(
+            f"invalid column type for {repr(key)}. {msg}: {repr(invalid_type)}"
+        )
+
+        self.key = key
+        self.invalid_type = invalid_type
 
 
 @dataclass(frozen=True)
@@ -355,12 +358,7 @@ class TableUpdate:
     Transactions are atomic. The entire transaction either commits or fails with an exception.
     """
 
-    # !!!
-    # WARNING: if you add more than one update type, the transaction will NOT BE ATOMIC
-    # we need to enable a postgraphile plugin to do mutations in transactions
-    # !!!
-
-    _record_upserts: List[_TableRecordsUpsertData] = field(
+    _record_mutations: List[_TableRecordsMutationData] = field(
         default_factory=list,
         init=False,
         repr=False,
@@ -369,6 +367,8 @@ class TableUpdate:
     )
 
     table: Table
+
+    # upsert record
 
     def upsert_record_raw_unsafe(
         self, *, name: str, values: Dict[str, DBValue]
@@ -392,10 +392,10 @@ class TableUpdate:
             name: Target record name.
             values: Column values that to set.
         """
-        self._record_upserts.append(_TableRecordsUpsertData(name, values))
+        self._record_mutations.append(_TableRecordsUpsertData(name, values))
 
     def upsert_record(self, name: str, **values: RegistryPythonValue) -> None:
-        """Update or create a record using.
+        """Update or create a record.
 
         A transport error will be thrown if non-existent columns are updated.
 
@@ -417,22 +417,72 @@ class TableUpdate:
 
             db_vals[k] = to_registry_literal(v, col.upstream_type["type"])
 
-        self._record_upserts.append(_TableRecordsUpsertData(name, db_vals))
+        self._record_mutations.append(_TableRecordsUpsertData(name, db_vals))
 
-    def _add_record_upserts_selection(self, updates: List[l.SelectionNode]) -> None:
-        if len(self._record_upserts) == 0:
+    def _add_record_upserts_selection(
+        self,
+        upserts: List[_TableRecordsUpsertData],
+        mutations: List[l.SelectionNode],
+        vars: Dict[str, Tuple[l.TypeNode, JsonValue]],
+    ) -> None:
+        if len(upserts) == 0:
             return
 
-        names: JsonValue = [x.name for x in self._record_upserts]
-        values: JsonValue = [
-            cast(Dict[str, JsonValue], x.values) for x in self._record_upserts
-        ]
+        names: _GqlJsonValue = [x.name for x in upserts]
+        values: JsonValue = [cast(Dict[str, JsonValue], x.values) for x in upserts]
 
-        res = _parse_selection("""
+        res = _parse_selection(
+            """
             catalogMultiUpsertSamples(input: {}) {
                 clientMutationId
             }
-        """)
+        """
+        )
+        assert isinstance(res, l.FieldNode)
+
+        argDataVar = f"upd{len(mutations)}ArgData"
+
+        args = l.ArgumentNode()
+        args.name = _name_node("input")
+        args.value = _json_value(
+            {
+                "argExperimentId": self.table.id,
+                "argNames": names,
+                "argData": _var_node(argDataVar),
+            }
+        )
+
+        res.alias = _name_node(f"upd{len(mutations)}")
+        res.arguments = tuple([args])
+
+        mutations.append(res)
+        vars[argDataVar] = (l.parse_type("[JSON]"), values)
+
+    # delete record
+
+    def delete_record(self, name: str) -> None:
+        """Delete a record.
+
+        Args:
+            name: Target record name.
+        """
+        self._record_mutations.append(_TableRecordsDeleteData(name))
+
+    def _add_record_deletes_selection(
+        self, deletes: List[_TableRecordsDeleteData], mutations: List[l.SelectionNode]
+    ) -> None:
+        if len(deletes) == 0:
+            return
+
+        names: _GqlJsonValue = [x.name for x in deletes]
+
+        res = _parse_selection(
+            """
+            catalogMultiDeleteSampleByName(input: {}) {
+                clientMutationId
+            }
+            """
+        )
         assert isinstance(res, l.FieldNode)
 
         args = l.ArgumentNode()
@@ -441,14 +491,174 @@ class TableUpdate:
             {
                 "argExperimentId": self.table.id,
                 "argNames": names,
-                "argData": values,
             }
         )
 
-        res.alias = _name_node(f"upd{len(updates)}")
+        res.alias = _name_node(f"upd{len(mutations)}")
         res.arguments = tuple([args])
 
-        updates.append(res)
+        mutations.append(res)
+
+    # upsert column
+
+    def upsert_column(
+        self,
+        key: str,
+        type: RegistryPythonType,
+        *,
+        required: bool = False,
+    ):
+        """Create a column. Support for updating columns is planned.
+
+        Args:
+            key: Identifier of the new column.
+            type:
+                Type of the new column.
+
+                Only a limited set of Python types is currently supported and
+                will be expanded over time.
+
+                :class:`latch.registry.types.RegistryPythonType` represents the currently supported types.
+            required:
+                If true, records without a value for this column are considered invalid.
+
+                Note that an explicit `None` value is different from a missing/empty value.
+                `None` is a valid value for an `Optional` (nullable) column marked as required.
+
+        """
+
+        registry_type: Optional[RegistryType] = None
+        if type is str:
+            registry_type = {"primitive": "string"}
+        if type is int:
+            registry_type = {"primitive": "integer"}
+        if type is float:
+            registry_type = {"primitive": "number"}
+        if type is date:
+            registry_type = {"primitive": "date"}
+        if type is datetime:
+            registry_type = {"primitive": "datetime"}
+        if type is bool:
+            registry_type = {"primitive": "boolean"}
+        if type is LatchFile:
+            registry_type = {"primitive": "blob"}
+        if type is LatchDir:
+            registry_type = {"primitive": "blob", "metadata": {"nodeType": "dir"}}
+
+        origin = get_origin(type)
+        if origin is not None:
+            if issubclass(origin, List):
+                inner_type = get_args(type)[0]
+                inner_type_origin = get_origin(inner_type)
+
+                if inner_type_origin is not None:
+                    if issubclass(inner_type_origin, LinkedRecordType):
+                        experiment_id = get_args(get_args(inner_type)[0])[0]
+                        registry_type = {
+                            "array": {
+                                "primitive": "link",
+                                "experimentId": experiment_id,
+                            }
+                        }
+                    else:
+                        raise InvalidColumnTypeError(
+                            key, type, "Unsupported list inner type"
+                        )
+                elif issubclass(inner_type, LatchFile):
+                    registry_type = {"array": {"primitive": "blob"}}
+                elif issubclass(inner_type, LatchDir):
+                    registry_type = {
+                        "array": {"primitive": "blob", "metadata": {"nodeType": "dir"}}
+                    }
+                else:
+                    raise InvalidColumnTypeError(
+                        key, type, "Unsupported list inner type"
+                    )
+
+            if issubclass(origin, LinkedRecordType):
+                experiment_id = get_args(get_args(type)[0])[0]
+                registry_type = {"primitive": "link", "experimentId": experiment_id}
+
+            if issubclass(origin, RegistryEnumDefinition):
+                members = list(get_args(t)[0] for t in get_args(get_args(type)[0]))
+                for x in members:
+                    if isinstance(x, str):
+                        continue
+                    raise InvalidColumnTypeError(
+                        key, type, f"Enum value {repr(x)} is not a string"
+                    )
+
+                registry_type = {
+                    "primitive": "enum",
+                    "members": members,
+                }
+
+        if isinstance(type, Enum):
+            members: List[str] = []
+            for f in cast(Type[Enum], type):
+                if not isinstance(f.value, str):
+                    raise InvalidColumnTypeError(
+                        key,
+                        type,
+                        (
+                            f"Enum value for {repr(f.name)} ({repr(f.value)}) is not a"
+                            " string"
+                        ),
+                    )
+
+                members.append(f.value)
+
+            registry_type = {
+                "primitive": "enum",
+                "members": members,
+            }
+
+        if registry_type is None:
+            raise InvalidColumnTypeError(key, type, "Unsupported type")
+
+        db_type: DBType = {"type": registry_type, "allowEmpty": not required}
+        self._record_mutations.append(_TableColumnUpsertData(key, db_type))
+
+    def _add_column_upserts_selection(
+        self,
+        upserts: List[_TableColumnUpsertData],
+        mutations: List[l.SelectionNode],
+        vars: Dict[str, Tuple[l.TypeNode, JsonValue]],
+    ) -> None:
+        if len(upserts) == 0:
+            return
+
+        keys: _GqlJsonValue = [x.key for x in upserts]
+        types: JsonValue = [cast(JsonValue, x.type) for x in upserts]
+
+        res = _parse_selection(
+            """
+            catalogExperimentColumnDefinitionMultiUpsert(input: {}) {
+                clientMutationId
+            }
+        """
+        )
+        assert isinstance(res, l.FieldNode)
+
+        argTypesVar = f"upd{len(mutations)}ArgTypes"
+
+        args = l.ArgumentNode()
+        args.name = _name_node("input")
+        args.value = _json_value(
+            {
+                "argExperimentId": self.table.id,
+                "argKeys": keys,
+                "argTypes": _var_node(argTypesVar),
+            }
+        )
+
+        res.alias = _name_node(f"upd{len(mutations)}")
+        res.arguments = tuple([args])
+
+        mutations.append(res)
+        vars[argTypesVar] = (l.parse_type("[JSON]!"), types)
+
+    # transaction
 
     def commit(self) -> None:
         """Commit this table update transaction.
@@ -459,21 +669,41 @@ class TableUpdate:
 
         Atomic. The entire transaction either commits or fails with an exception.
         """
-        updates: List[l.SelectionNode] = []
+        mutations: List[l.SelectionNode] = []
+        vars: Dict[str, Tuple[l.TypeNode, JsonValue]] = {}
 
-        self._add_record_upserts_selection(updates)
-
-        if len(updates) == 0:
+        if len(self._record_mutations) == 0:
             return
 
-        sel_set = l.SelectionSetNode()
-        sel_set.selections = tuple(updates)
+        def _add_record_data_selection(cur):
+            if isinstance(cur[0], _TableRecordsUpsertData):
+                self._add_record_upserts_selection(cur, mutations, vars)
+            if isinstance(cur[0], _TableRecordsDeleteData):
+                self._add_record_deletes_selection(cur, mutations)
+            if isinstance(cur[0], _TableColumnUpsertData):
+                self._add_column_upserts_selection(cur, mutations, vars)
 
-        doc = l.parse("""
+        cur = [self._record_mutations[0]]
+        for mut in self._record_mutations[1:]:
+            if isinstance(mut, type(cur[0])):
+                cur.append(mut)
+                continue
+
+            _add_record_data_selection(cur)
+            cur = [mut]
+
+        _add_record_data_selection(cur)
+
+        sel_set = l.SelectionSetNode()
+        sel_set.selections = tuple(mutations)
+
+        doc = l.parse(
+            """
             mutation TableUpdate {
                 placeholder
             }
-        """)
+        """
+        )
 
         assert len(doc.definitions) == 1
         mut = doc.definitions[0]
@@ -481,7 +711,13 @@ class TableUpdate:
         assert isinstance(mut, l.OperationDefinitionNode)
         mut.selection_set = sel_set
 
-        execute(doc)
+        mut.variable_definitions = tuple(
+            _var_def_node(k, t) for k, (t, _) in vars.items()
+        )
+
+        # todo(maximsmol): catch errors here and raise appropriate Python exceptions
+        # 1. column upsert: already exists
+        execute(doc, {k: v for k, (_, v) in vars.items()})
 
         self.clear()
 
@@ -490,4 +726,4 @@ class TableUpdate:
 
         May be called to cancel any pending updates that have not been committed.
         """
-        self._record_upserts.clear()
+        self._record_mutations.clear()